+++ examples/demoCubes.cpp
@@ -1,10 +1,23 @@
-/**
- * @file    demoCubes.cpp
- * @brief
- * @author  Juan Guerrero Martín <juan.guerrero@urjc.es>
- * @date    2016
- * @remarks Copyright (c) 2016 GMRV/URJC. All rights reserved.
- * Do not distribute without further notice.
+/*
+ * Copyright (c) 2014-2016 GMRV/URJC.
+ *
+ * Authors: Juan Guerrero Martín
+ *
+ * This file is part of ReTo <https://gitlab.gmrv.es/nsviz/ReTo>
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License version 3.0 as published
+ * by the Free Software Foundation.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE.  See the GNU Lesser General Public License for more
+ * details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with this library; if not, write to the Free Software Foundation, Inc.,
+ * 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
+ *
  */
 
 // std.
@@ -64,7 +77,7 @@ float currentYaw = 90.0f;
 float currentPitch = 0.0f;
 /**/
 
-std::vector< float > matrix4fToVector16f( Eigen::Matrix4f inputMatrix );
+std::vector< float > matrix4fToVector16f( const Eigen::Matrix4f& inputMatrix );
 
 void renderFunc( void );
 void resizeFunc( int width, int height );
@@ -173,7 +186,7 @@ void destroy( void )
 {
 }
 
-std::vector< float > matrix4fToVector16f( Eigen::Matrix4f inputMatrix )
+std::vector< float > matrix4fToVector16f( const Eigen::Matrix4f& inputMatrix )
 {
   std::vector< float > toReturn;