+++ Changelog.md
@@ -1,6 +1,7 @@
 # ReTo - Render auxiliary Tools
 
 ## Changelog
+- (2017/04/04) Camera redesign [!32].
 - (2016/09/15) Added Spline navigation.
 - (2016/09/16) Added OBJ Loader.
 - (2016/10/05) Fixed picking algorithm.